+++ src/app/utils/cluster-storage/generic-fields.js
@@ -19,6 +19,12 @@ export default [
     tip: true,
   },
   {
+    name: 'readonly',
+    type: 'checkbox',
+    optional: true,
+    tip: true,
+  },
+  {
     name: 'importedStorage',
     type: 'checkbox',
     defaultValue: false,